add vehicule system

+++ OpenNos.DAL.EF.MySQL/DB/Item.cs
@@ -26,7 +26,7 @@ namespace OpenNos.DAL.EF.MySQL.DB
         public string Name { get; set; }
         public byte ItemType { get; set; }
         public byte EquipmentSlot { get; set; }
-        public byte Morph { get; set; }
+        public short Morph { get; set; }
         public byte Type { get; set; }
         public byte Classe { get; set; }
         public byte Blocked { get; set; }
